pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cagey"
version = "0.1.0"
description = "Analyze colleagues' and bosses' work chats for sentiment and sub-auditions using Claude."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "Cagey Project" }]
dependencies = [
    "anthropic>=0.40.0",
    "typer>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "plotly>=5.22.0",
    "python-dateutil>=2.9.0",
    "tenacity>=8.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
]

[project.scripts]
cagey = "cagey.cli:app"

[tool.setuptools.packages.find]
include = ["cagey*"]
exclude = ["tests*", "examples*"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
